rename: install_wasm_target.rs
